Bing no longer returns tracking query string params (#5)
Previously, URLs were hosted on bing.com and redirected via query params. These params included `h` which this library considered an ID and used it to strip duplicate results (since there can be overlap when paginating).

The API appears to be returning source URLs now, so we can directly strip duplicates that way w/o having to rely on query params to determine an ID. I also simplified some of the surrounding logic while I was in here.